backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
Flask-Talisman==1.1.0
celery==5.3.4
redis==5.0.1
openai==1.3.7
python-dotenv==1.0.0
psycopg2-binary==2.9.7
PyPDF2==3.0.1
python-docx==1.1.0
pytest==7.4.3
pytest-cov==4.1.0
gunicorn==21.2.0
Werkzeug==2.3.7
email-validator==2.0.0
bcrypt==4.0.1 